feat: improve directory and naming convention
