+++ src/component/connection-status/connection-status.module.scss
