+++ src/component/multi-chain-wallet-adapters/connect-task-bar.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 import PropTypes from 'prop-types';
-import TaskBox from '../task-box/task-box';
+import TaskBox from '../connection-box/task-box';
 import styles from './wallet-button.module.scss';
 import { WalletMultiButton } from '@solana/wallet-adapter-react-ui';
 import { ConnectWalletTaskBoxProps } from './../../types/wallet-types';
